move code notes into separate model
