+++ src/RA_Integration.vcxproj
@@ -62,6 +62,7 @@
     <ClCompile Include="data\models\AchievementModel.cpp" />
     <ClCompile Include="data\models\AssetModelBase.cpp" />
     <ClCompile Include="data\models\CapturedTriggerHits.cpp" />
+    <ClCompile Include="data\models\CodeNotesModel.cpp" />
     <ClCompile Include="data\models\LeaderboardModel.cpp" />
     <ClCompile Include="data\models\LocalBadgesModel.cpp" />
     <ClCompile Include="data\models\RichPresenceModel.cpp" />
@@ -216,6 +217,7 @@
     <ClInclude Include="data\models\AchievementModel.hh" />
     <ClInclude Include="data\models\AssetModelBase.hh" />
     <ClInclude Include="data\models\CapturedTriggerHits.hh" />
+    <ClInclude Include="data\models\CodeNotesModel.hh" />
     <ClInclude Include="data\models\LeaderboardModel.hh" />
     <ClInclude Include="data\models\LocalBadgesModel.hh" />
     <ClInclude Include="data\models\RichPresenceModel.hh" />
