+++ src/data/context/GameContext.hh
@@ -102,123 +102,6 @@ public:
     /// </summary>
     void RefreshUnlocks() { RefreshUnlocks(false, 0); }
 
-    /// <summary>
-    /// Returns the note associated with the specified address.
-    /// </summary>    
-    /// <returns>The note associated to the address, <c>nullptr</c> if no note is associated to the address.</returns>
-    const std::wstring* FindCodeNote(ra::ByteAddress nAddress) const
-    {
-        const auto pIter = m_mCodeNotes.find(nAddress);
-        return (pIter != m_mCodeNotes.end()) ? &pIter->second.Note : nullptr;
-    }
-
-    /// <summary>
-    /// Returns the address of the first byte containing the specified code note.
-    /// </summary>
-    /// <returns>
-    ///  Returns 0xFFFFFFFF if not found.
-    /// </returns>
-    ra::ByteAddress FindCodeNoteStart(ra::ByteAddress nAddress) const;
-
-    /// <summary>
-    /// Returns the note associated with the specified address.
-    /// </summary>
-    /// <returns>
-    ///  The note associated to the address or neighboring addresses based on <paramref name="nSize" />.
-    ///  Returns an empty string if no note is associated to the address.
-    /// </returns>
-    std::wstring FindCodeNote(ra::ByteAddress nAddress, MemSize nSize) const;
-
-    /// <summary>
-    /// Returns the note associated with the specified address.
-    /// </summary>
-    /// <param name="nAddress">The address to look up.</param>
-    /// <param name="sAuthor">The author associated to the address.</param>
-    /// <returns>The note associated to the address, <c>nullptr</c> if no note is associated to the address.</returns>
-    const std::wstring* FindCodeNote(ra::ByteAddress nAddress, _Inout_ std::string& sAuthor) const
-    {
-        const auto pIter = m_mCodeNotes.find(nAddress);
-        if (pIter == m_mCodeNotes.end())
-            return nullptr;
-
-        sAuthor = pIter->second.Author;
-        return &pIter->second.Note;
-    }
-
-    /// <summary>
-    /// Returns the number of bytes associated to the code note at the specified address.
-    /// </summary>
-    /// <param name="nAddress">Address to query.</param>
-    /// <returns>Number of bytes associated to the code note, or 0 if no note exists at the address.</returns>
-    /// <remarks>Only works for the first byte of a multi-byte address.</remarks>
-    unsigned GetCodeNoteBytes(ra::ByteAddress nAddress) const
-    {
-        const auto pIter = m_mCodeNotes.find(nAddress);
-        return (pIter == m_mCodeNotes.end()) ? 0 : pIter->second.Bytes;
-    }
-
-    /// <summary>
-    /// Returns the number of bytes associated to the code note at the specified address.
-    /// </summary>
-    /// <param name="nAddress">Address to query.</param>
-    /// <returns>Number of bytes associated to the code note, or 0 if no note exists at the address.</returns>
-    /// <remarks>Only works for the first byte of a multi-byte address.</remarks>
-    MemSize GetCodeNoteMemSize(ra::ByteAddress nAddress) const
-    {
-        const auto pIter = m_mCodeNotes.find(nAddress);
-        return (pIter == m_mCodeNotes.end()) ? MemSize::Unknown : pIter->second.MemSize;
-    }
-
-    /// <summary>
-    /// Enumerates the code notes
-    /// </summary>
-    /// <remarks>
-    /// <paramref name="callback" /> is called for each known code note. If it returns <c>false</c> enumeration stops.
-    /// </remarks>
-    void EnumerateCodeNotes(std::function<bool(ra::ByteAddress nAddress)> callback) const
-    {
-        for (auto& pCodeNote : m_mCodeNotes)
-        {
-            if (!callback(pCodeNote.first))
-                break;
-        }
-    }
-
-    /// <summary>
-    /// Enumerates the code notes
-    /// </summary>
-    /// <remarks>
-    /// <paramref name="callback" /> is called for each known code note. If it returns <c>false</c> enumeration stops.
-    /// </remarks>
-    void EnumerateCodeNotes(std::function<bool(ra::ByteAddress nAddress, unsigned int nBytes, const std::wstring& sNote)> callback) const
-    {
-        for (auto& pCodeNote : m_mCodeNotes)
-        {
-            if (!callback(pCodeNote.first, pCodeNote.second.Bytes, pCodeNote.second.Note))
-                break;
-        }
-    }
-
-    /// <summary>
-    /// Sets the note to associate with the specified address.
-    /// </summary>
-    /// <param name="nAddress">The address to set the note for.</param>
-    /// <param name="sNote">The new note for the address.</param>
-    /// <returns><c>true</c> if the note was updated, </c>false</c> if an error occurred.</returns>
-    virtual bool SetCodeNote(ra::ByteAddress nAddress, const std::wstring& sNote);
-
-    /// <summary>
-    /// Deletes the note associated with the specified address.
-    /// </summary>
-    /// <param name="nAddress">The address to delete the note for.</param>
-    /// <returns><c>true</c> if the note was deleted, </c>false</c> if an error occurred.</returns>
-    virtual bool DeleteCodeNote(ra::ByteAddress nAddress);
-
-    /// <summary>
-    /// Returns the number of known code notes
-    /// </summary>
-    size_t CodeNoteCount() const noexcept { return m_mCodeNotes.size(); }
-
     class NotifyTarget
     {
     public:
@@ -246,8 +129,6 @@ private:
 protected:
     void RefreshUnlocks(bool bUnpause, int nPopup);
     void UpdateUnlocks(const std::set<unsigned int>& vUnlockedAchievements, bool bUnpause, int nPopup);
-    void RefreshCodeNotes();
-    void AddCodeNote(ra::ByteAddress nAddress, const std::string& sAuthor, const std::wstring& sNote);
     void ShowSimplifiedScoreboard(ra::LeaderboardID nLeaderboardId, int nScore) const;
 
     void OnActiveGameChanged();
@@ -261,17 +142,7 @@ protected:
     std::string m_sGameImage;
     Mode m_nMode{};
 
-    struct CodeNote
-    {
-        std::string Author;
-        std::wstring Note;
-        unsigned int Bytes = 1;
-        MemSize MemSize = MemSize::Unknown;
-    };
-    std::map<ra::ByteAddress, CodeNote> m_mCodeNotes;
-
 private:
-    static void ExtractSize(CodeNote& pNote);
     void CheckForMastery();
 
     /// <summary>
