+++ src/data/context/SessionTracker.cpp
@@ -257,7 +257,7 @@ std::wstring SessionTracker::GetCurrentActivity() const
             const auto* pAsset = pGameContext.Assets().GetItemAt(i);
             if (pAsset)
             {
-                if (pAsset->GetType() == ra::data::models::AssetType::LocalBadges)
+                if (pAsset->IsHidden())
                     continue;
 
                 if (pAsset->GetCategory() == ra::data::models::AssetCategory::Local)
